requirements.txt
gradio
pandas
matplotlib
seaborn
pillow
pydantic-ai

